requirements.txt
streamlit
groq
google-generativeai
python-dotenv
plotly
requests